Add a Maven extension version
If a user refers to the properties generated by the Maven plugin in a
project dependency, the properties are not interpolated and thus the
referring dependencies are not resolved.

To work around this problem, this plugin must be implemented as an
extension, and the extension must perform the necessary interpolation
before it enters the 'validate' phase.